refactor(web): remove dead code, prune deps, split HeroSection, normalize naming
- Delete 6 orphaned files (animated-mobile-nav/search/sidebar, progress-bar,
  use-mobile, use-toast) plus the unused styles/globals.css
- Prune ~30 unused @radix-ui/* packages and ~18 other unused libs (recharts,
  zod, cmdk, vaul, sonner, react-hook-form, date-fns, etc.) and bogus fs/path
  entries; refresh yarn.lock
- Split 650-line HeroSection into a thin orchestrator + components/hero/
  (hero-tools, use-tool-animation hook, hero-intro, tool-selector, labeling-canvas)
- Rename PascalCase component files to kebab-case to match house style
