+++ apps/web/app/blogs/page.tsx
@@ -2,7 +2,7 @@ import { promises as fs } from "fs"
 import path from "path"
 import matter from "gray-matter"
 import Link from "next/link"
-import BlogSidebar from "./BlogSidebar"
+import BlogSidebar from "./blog-sidebar"
 
 interface Blog {
   title: string
