+++ apps/web/app/docs/page.tsx
@@ -1,4 +1,4 @@
-import DocumentationClientPage from "./DocumentationClientPage"
+import DocumentationClientPage from "./documentation-client-page"
 export default async function DocumentationPage() {
   return <DocumentationClientPage />
 }
