+++ apps/web/components/feature-highlights.tsx
