+++ apps/web/components/footer.tsx
